requirements.txt
pyaudio
aiogram
numpy
scipy
audioop-lts; python_version >= "3.13"
python-dotenv
